+++ lib/libchen/CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file(see http://k
 ### Fixed
 - ini: fix parse error if comment between two properties.
 - str: fix replace string didn't consider the all param.
+- fs: fix method absolute didn't remove ".." if path is "../a/b".
 
 ## [1.0.1] - 2016-06-21
 ### Added